# docs/openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: TonForge Agent API
  version: "1.0.0"
  summary: Machine-facing API for AI agents to automate a seller's product lifecycle.
  description: |
    The **Agent API** lets an autonomous agent (or any HTTP client acting on a
    seller's behalf) manage that seller's marketplace listings, distribution
    manifests, and read their orders — without ever touching a browser session.

    ## Authentication
    Every request authenticates with a **Personal Access Token (PAT)** issued by
    a *verified* seller. Tokens look like `tfa_XXXXXXXX…` and are passed either as:

    - `Authorization: Bearer tfa_…`  (preferred, standard), or
    - `X-Agent-Token: tfa_…`         (fallback for clients that reserve `Authorization`).

    The acting wallet is read **from the token**, never from the request body or
    a header. An agent therefore cannot act for a seller other than the one who
    issued its token, no matter what it puts in the payload.

    Tokens are issued, listed, and revoked by the human seller through the
    session-authenticated `agent-tokens` endpoints (see the `Token management`
    tag). The plaintext token is shown **exactly once** at issue time.

    ## Authorization (scopes)
    A token carries a set of scopes. `listings:write` implies `listings:read`;
    `distribution:write` implies `listings:read`. A request missing a required
    scope returns `403 SCOPE_FORBIDDEN`.

    ## Hard gates (defence in depth)
...
    call time:
...
    own progress.

...
servers:
...
security:
  - agentBearer: []
  - agentHeader: []
paths:
...
  /listings/{id}/distribution:
    put:
      tags: [Distribution]
      operationId: setDistribution
      summary: Set the distribution manifest
      description: |
        Attaches a distribution manifest (an R2 object or a GitHub release asset)
        to a listing you own and moves it to the `draft` distribution state.
        Requires `distribution:write`. Run `…/distribution/verify` afterward to
        confirm the artifact resolves and matches its declared sha256.
      security:
        - agentBearer: [distribution:write]
        - agentHeader: [distribution:write]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest]
              properties:
                manifest:
                  $ref: "#/components/schemas/DistributionManifest"
                ttlSec:
                  type: integer
                  description: Signed-URL TTL in seconds for resolved downloads.
                  default: 3600
      responses:
        "200":
          description: Manifest stored.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      ok: { type: boolean }
                      state: { type: string, example: draft }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotOwner" }
  /listings/{id}/distribution/verify:
    post:
      tags: [Distribution]
      operationId: verifyDistribution
      summary: Verify the distribution manifest
      description: |
        Resolves the previously-set manifest, hashes the artifact, and compares it
        to the declared sha256. On match the listing moves to `verified`; on
        mismatch to `manifest_drift`. Requires `distribution:write`.
      security:
        - agentBearer: [distribution:write]
        - agentHeader: [distribution:write]
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Verification result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      matches: { type: boolean }
                      sha256: { type: string }
                      size: { type: integer }
                      state: { type: string, enum: [verified, manifest_drift] }
        "400":
          description: No manifest has been set yet (`NO_MANIFEST`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotOwner" }
...
  /buyer/listings/{id}/download:
    get:
      tags: [Agentic buying]
      operationId: buyerDownload
      summary: "Buy step 3: signed download URL for a purchased listing"
      description: >-
        Returns a short-lived signed URL plus the expected sha256. Gated on the
        minted license NFT for the buyer wallet and a clean antivirus verdict;
        ≤20 URLs/day per purchase. Verify the sha256 after download.
      security:
        - agentToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Signed URL issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      url: { type: string }
                      expiresInSec: { type: integer }
                      sha256: { type: string }
        "403": { description: No entitlement / license not minted yet. }
        "404": { description: Listing or verified build not found. }
        "429": { description: Download rate limit (20/day). }
components:
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: tfa
      description: "`Authorization: Bearer tfa_…`"
    agentHeader:
      type: apiKey
      in: header
      name: X-Agent-Token
      description: "`X-Agent-Token: tfa_…` — fallback channel."
    sessionJwt:
      type: http
      scheme: bearer
      description: Seller user-session JWT. Used only by the Token management endpoints.
  parameters:
    ListingId:
      name: id
      in: path
      required: true
      description: Listing document id.
      schema: { type: string }
  schemas:
...
    DistributionManifest:
      type: object
      description: >-
        An R2 object or GitHub release asset descriptor with an expected sha256.
        See `docs/byos-distribution.md` for the full field set.
      additionalProperties: true
      properties:
        kind: { type: string, enum: [r2, github] }
        sha256: { type: string }
        filename: { type: string }
...
    Error:
      type: object
      description: Route-level error shape.
      properties:
        error: { type: string }
        code: { type: string }
    AuthError:
      type: object
      description: Auth-middleware error shape.
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        code: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
          examples:
            noToken: { value: { success: false, message: "Agent token required", code: "NO_AGENT_TOKEN" } }
            badToken: { value: { success: false, message: "Invalid or expired token", code: "BAD_AGENT_TOKEN" } }
...
    NotOwner:
      description: The listing exists but is owned by a different wallet.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Not your listing", code: "NOT_OWNER" }
...
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Missing required fields", code: "VALIDATION" }
...